Content: gate staging
